Rework screen to match latest Jesse design
