+++ apps/public-reference/pages/applications/preferences/general.tsx
@@ -41,7 +41,7 @@ export default () => {
       <FormCard>
         <p className="text-bold">
           <strong>
-            <Link href="/applications/preferences/live-or-work">
+            <Link href="/applications/preferences/select">
               <a>{t("t.back")}</a>
             </Link>
           </strong>
